all commit 3

+++ Object.pde
@@ -1,5 +1,6 @@
 abstract class Obj{
   float xPos, yPos, dia;
+  float callTime = millis();
   Figure figure;
   Move move;
   boolean delete;
@@ -26,6 +27,18 @@ abstract class Obj{
     this.delete = true;
   }
 
+  float getTime(){
+    return (millis() - this.callTime) / 1000.0;
+  }
+
+  void evaporate(float speed){
+    this.dia -= speed;
+  }
+
+  boolean catchDelete(){
+    return (catchStop() || catchDiaZero() || catchOutOfScreen());
+  }
+
   boolean catchStop(){
     if(this.move.vector2D.speed() <= 0.1)
       return true;
@@ -44,6 +57,7 @@ abstract class Obj{
   }
 
   boolean catchCollision(Obj obj){
+    if(obj == this) return false;
     if(dist(this.xPos, this.yPos, obj.xPos, obj.yPos) < obj.dia/2 + this.dia/2)
       return true;
     return false;
@@ -67,43 +81,89 @@ void objUpdate(Object arrayList){
 
 class BulletObj extends Obj{
   float damage;
+  boolean myBullet = false;
   BulletObj(float xPos, float yPos, float dia, Figure figure, Move move){
     super(xPos, yPos, dia, figure, move);
-    this.damage = 1;
+    this.damage = 5;
   }
+  BulletObj(){}
   BulletObj(float xPos, float yPos, float dia, Figure figure, Move move, float damage){
     super(xPos, yPos, dia, figure, move);
     this.damage = damage;
   }
   boolean update(){
-    super.figure.update(super.xPos, super.yPos, super.dia);
-    for(Obj obj: enemys)
-      if(this.catchCollision(obj))
-        for(int i = 0; i < 5; i++){
-          fireSample(super.xPos, super.yPos, 30, new color[]{color(0, 100, 255, 100)});
-          fireSample(super.xPos, super.yPos, 90, new color[]{color(255, 0, 0, 100)});
+    if(this.playerHit()){
+      this.myBullet = true;
+    }
+    this.figureUpdate();
+    if(this.myBullet){
+      super.figure.toMyBullet();
+      for(EnemyObj obj: enemys)
+        if(this.catchCollision(obj)){
+          obj.damage(this.damage);
+          for(int i = 0; i < 5; i++)
+            fireSample(super.xPos, super.yPos, this.dia, figure.col);
           super.delete();
         }
+    }
     super.setMoveState(super.move.action(super.getMoveState()));
-    if(super.catchOutOfScreen())
+    if(super.catchDelete())
       super.delete();
     return super.delete;
+  }
+
+  private boolean playerHit(){
+    for(BarObj player: bars)
+      if(player.catchHit(this)){
+        this.adjustBound(player);
+        return true;
+      }
+    return false;
+  }
+
+  private void normalBound(){
+    super.move.vector2D.yReflect();
+  }
+
+  private void adjustBound(Obj player){
+    super.move.vector2D.angle(degrees(atan2(super.yPos - player.yPos, super.xPos - player.xPos)));
+  }
+
+  private void randomBound(){
+    if(super.move.vector2D.ySpeed() < 0)
+      super.move.vector2D.angle(random(210, 330));
+    else
+      super.move.vector2D.angle(random(30, 150));
+  }
+
+  void figureUpdate(){
+    pushMatrix();
+    translate(super.xPos, super.yPos);
+    rotate(atan2(super.move.vector2D.ySpeed(), super.move.vector2D.xSpeed()) + radians(90));
+    super.figure.update(0, 0, super.dia);
+    popMatrix();
   }
 }
 
 class EnemyObj extends Obj{
-  float hp, hpBar;
+  float hp, hpBar, maxHp;
+  EnemyObj(){}
   EnemyObj(float xPos, float yPos, float dia, Figure figure, Move move){
     super(xPos, yPos, dia, figure, move);
     this.hp = 5;
     this.hpBar = dia / this.hp;
+    this.maxHp = hp;
   }
   EnemyObj(float xPos, float yPos, float dia, Figure figure, Move move, float hp){
     super(xPos, yPos, dia, figure, move);
     this.hp = hp;
     this.hpBar = dia / this.hp;
+    this.maxHp = hp;
   }
   boolean update(){
+    /* for(Obj obj: enemys){ */
+    /*   if(catchCollision(obj)) super.move.vector2D.rotation(180); */
+    /* } */
     super.figure.update(super.xPos, super.yPos, super.dia);
     this.drawLife();
     super.setMoveState(super.move.action(super.getMoveState()));
@@ -113,7 +173,17 @@ class EnemyObj extends Obj{
   }
   void drawLife(){
     fill(255, 170);
-    rect(super.xPos - super.dia/2, super.yPos + super.dia/2, super.dia, 10);
+    rect(super.xPos - super.dia/2, super.yPos + super.dia/2, this.hpBar * this.hp, 10);
+  }
+  void damage(float damage){
+    this.hp -= damage;
+    if(this.hp <= 0){
+      this.hp = 0;
+      inf.addScore(this.maxHp);
+      for(int i = 0; i < 5; i++)
+        fireSample(super.xPos, super.yPos, this.dia, figure.col);
+      super.delete();
+    }
   }
 }
 
@@ -131,8 +201,38 @@ class ParticleObj extends Obj{
   }
 }
 
-abstract class BarObj extends Obj{
-  BarObj(float xPos, float yPos, float dia, Figure figure, Move move){
+class BarObj extends Obj{
+  BarObj(){}
+  BarObj(float xPos, float yPos, float dia, BarFigure figure){
+    super(xPos, yPos, dia, figure, stopMove());
+  }
+  BarObj(float xPos, float yPos, float dia, BarFigure figure, Move move){
     super(xPos, yPos, dia, figure, move);
+  }
+  boolean update(){
+    super.figure.update(super.xPos, super.yPos, super.dia);
+    return super.delete;
+  }
+
+  void leftDisplace(float speed){
+    super.xPos -= speed;
+    if(super.xPos < dia/2) super.xPos = dia/2;
+  }
+
+  void rightDisplace(float speed){
+    super.xPos += speed;
+    if(width - super.dia/2 < super.xPos) super.xPos = width - super.dia/2;
+  }
+
+  boolean catchHit(Obj obj){
+    if(obj.yPos + obj.dia/2 + obj.move.vector2D.ySpeed() > super.yPos - 5 && 
+        obj.yPos - obj.dia/2 + obj.move.vector2D.ySpeed() < super.yPos + 5 && 
+        obj.xPos > super.xPos - super.dia/2 && 
+        obj.xPos < super.xPos + super.dia/2){
+      if(0 <= obj.move.vector2D.ySpeed()) obj.yPos = super.yPos - 10;
+      else obj.yPos = super.yPos + 10;
+      return true;
+    }
+    return false;
   }
 }
